Fragments and Tablet UI added
Original program is still there. To get to the purposed layout, hit Test
or Test 2 in the main menu. The main purpose is to showcase the Tablet
UI. Hence, these tests do not have full menu functionality.
Needs a lot of testings but here is the run down:

- All the novel listings + details have been changed to fragments
- UI layout + animations have been implemented
- Synopsis fragment has a purpose >.>

OMISSIONS:
- Some menu buttons have been deleted out in the test. This is due to
potential confusion to the user.